+++ response/item_test.go
@@ -23,7 +23,7 @@ func (suite *ResponseItemTestSuite) TestDefaultMessage() {
 	suite.resExpected.Message = "Data retrieval successfully"
 	status, response := Item(suite.item, "")
 	assert.Equal(suite.T(), suite.resExpected, response)
-	assert.Equal(suite.T(), int16(200), status)
+	assert.Equal(suite.T(), 200, status)
 }
 
 func (suite *ResponseItemTestSuite) TestCustomMessage() {
@@ -32,7 +32,7 @@ func (suite *ResponseItemTestSuite) TestCustomMessage() {
 	suite.resExpected.Message = message
 	status, response := Item(suite.item, message)
 	assert.Equal(suite.T(), suite.resExpected, response)
-	assert.Equal(suite.T(), int16(200), status)
+	assert.Equal(suite.T(), 200, status)
 }
 
 func TestResponseItemTestSuite(t *testing.T) {
