+++ response/unauthorized.go
@@ -1,6 +1,6 @@
 package response
 
-func Unauthorized(message string) (status int16, res commonResponse) {
+func Unauthorized(message string) (status int, res commonResponse) {
 	meg := "Unauthorized"
 	if message != "" {
 		meg = message
